requirements.txt
anthropic>=0.39.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pypdf>=4.0.0
cryptography>=41.0.0
rank-bm25>=0.2.2
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
